cw2_chatbot_task1_task2_tts/requirements.txt
flask
flask_pymongo
torch
nltk
nltk.stem.porter
experta
datetime
selenium
webdriver_manager
frozendict